refactor tests and add stubs for missing tests
