+++ tests/its_live_monitoring/test_itslive.py
@@ -1,0 +1,19 @@
+def test_get_datetime():
+    # TODO
+    pass
+
+def test_sort_earliest_first():
+    # TODO
+    pass
+
+def test_bursts_in_item():
+    # TODO
+    pass
+
+def test_pair_exists():
+    # TODO
+    pass
+
+def test_deduplicate_published_pairs():
+    # TODO
+    pass
